T05 : change price to Double + navigation from register to profile + fixed display profil

+++ app/src/main/java/c0d3/vitreen/app/fragments/adding/Adding2Fragment.kt
@@ -100,7 +100,7 @@ class Adding2Fragment : VFragment(
                     val product = Product(
                         title = title,
                         description = description,
-                        price = price.toLong(),
+                        price = price.toDouble(),
                         brand = editTextBrand.text.toString(),
                         size = editTextDimensions.text.toString(),
                         locationId = locationId,
